Conflicts:
	projetoFinal/src/main/java/com/senai/projetofinal/dao/AgendaDAO.java
	projetoFinal/src/main/java/com/senai/projetofinal/dao/CategoriaDAO.java
	projetoFinal/src/main/java/com/senai/projetofinal/dao/ClienteDAO.java
	projetoFinal/src/main/java/com/senai/projetofinal/dao/EnderecoDAO.java
	projetoFinal/src/main/java/com/senai/projetofinal/dao/ServicoDAO.java

+++ projetoFinal/src/main/java/com/senai/projetofinal/dao/ClienteDAO.java
@@ -28,7 +28,7 @@ public class ClienteDAO {
         em.persist(cliente);
     }
 
-    public void excluir(Long id) {
+    public void excluir(Integer id) {
         em.remove(em.getReference(Cliente.class, id));
     }
